requirements.txt
numpy
olefile
opencv-python
Pillow
qrcode
six
zbarlight
